Isnull check for referencedOrderItem
This was causing an issue when return orders weren't created from an original order - eg: an exchange from a customer who bought through a reseller.  The filter line on the accountContentAccessSmartList was throwing an error, so the return was receivable

+++ model/service/OrderService.cfc
@@ -2356,13 +2356,15 @@ component extends="HibachiService" persistent="false" accessors="true" output="f
 			}
 
 			// If there was one or more accountContentAccess associated with the referenced orderItem then we need to remove them.
-			var accountContentAccessSmartList = getAccountService().getAccountContentAccessSmartList();
-			accountContentAccessSmartList.addFilter("OrderItem.orderItemID", stockReceiverItem.getOrderItem().getReferencedOrderItem().getOrderItemID());
-			var accountContentAccesses = accountContentAccessSmartList.getRecords();
-			for (var accountContentAccess in accountContentAccesses){
+			if(!isnull(stockReceiverItem.getOrderItem().getReferencedOrderItem())){
+				var accountContentAccessSmartList = getAccountService().getAccountContentAccessSmartList();
+				accountContentAccessSmartList.addFilter("OrderItem.orderItemID", stockReceiverItem.getOrderItem().getReferencedOrderItem().getOrderItemID());
+				var accountContentAccesses = accountContentAccessSmartList.getRecords();
+				for (var accountContentAccess in accountContentAccesses){
 
-    			getAccountService().deleteAccountContentAccess( accountContentAccess );
+				getAccountService().deleteAccountContentAccess( accountContentAccess );
 
+				}
 			}
 		}
